Revise reputation documentation for clarity and detail
Updated the reputation documentation to clarify the role of on-chain transactions and off-chain payment promises in the Chiral Network. Added details on the payment handshake process and non-payment complaint lifecycle. Also added an Asymmetric Protection where we treat the downloader as the victim in unsolved malicious cases, as recommended by professor